Move filter functions into reader
